first draft of featuredCard
