+++ src/components/FeaturedCard/components/FeaturedRest.js
@@ -1,61 +1,31 @@
 import React from 'react';
-import { Table, Icon } from 'antd';
+import { Card } from 'antd';
+import PropTypes from 'prop-types';
 
-const columns = [
-  {
-    title: 'Name',
-    dataIndex: 'name',
-    key: 'name',
-    render: text => <a href="#">{text}</a>
-  },
-  {
-    title: 'Age',
-    dataIndex: 'age',
-    key: 'age'
-  },
-  {
-    title: 'Address',
-    dataIndex: 'address',
-    key: 'address'
-  },
-  {
-    title: 'Action',
-    key: 'action',
-    render: (text, record) => (
-      <span>
-        <a href="#">Action 一 {record.name}</a>
-        <span className="ant-divider" />
-        <a href="#">Delete</a>
-        <span className="ant-divider" />
-        <a href="#" className="ant-dropdown-link">
-          More actions <Icon type="down" />
-        </a>
-      </span>
-    )
-  }
-];
+const gridStyle = {
+  width: '25%',
+  textAlign: 'center'
+};
 
-const data = [
-  {
-    key: '1',
-    name: 'John Brown',
-    age: 32,
-    address: 'New York No. 1 Lake Park'
-  },
-  {
-    key: '2',
-    name: 'Jim Green',
-    age: 42,
-    address: 'London No. 1 Lake Park'
-  },
-  {
-    key: '3',
-    name: 'Joe Black',
-    age: 32,
-    address: 'Sidney No. 1 Lake Park'
-  }
-];
-
-const FeaturedRest = () => <Table columns={columns} dataSource={data} />;
+const FeaturedRest = ({ data }) => (
+  <Card title="Card Title" noHovering>
+    {data.map(item => <Card.Grid style={gridStyle}>{item.title}</Card.Grid>)}
+  </Card>
+);
 
 export default FeaturedRest;
+
+FeaturedRest.propTypes = {
+  data: PropTypes.arrayOf(Object)
+};
+
+FeaturedRest.defaultProps = {
+  data: [
+    {
+      image: 'https://www.equalserving.com/images/es1scplugin/image-not-available.png',
+      title: 'Title not available',
+      text: 'Text not available',
+      action: '🚫'
+    }
+  ]
+};
